fix: rename component files
